Project as interface #1053
